Changing "in"-operator to "case-of"-block for SVG parsing.
Delphi's DCC64 compiler is notoriously bad when it comes to optimizing the "in" operator (tested with Delphi XE2-D11). This patch replaces all the "in"-operators with "case-of" blocks that even DCC64 can optimize. FPC generates equally optimized code for "in" and "case-of".

Delphi also has problems inlining functions if they are used in "if" or "while" conditions. This is also addressed by moving the conditions into the while-loop's body.

CharInSet, even inlined, uses the relatively slow "bt" opcode. Thus most of those calls are replaced by "case-of".