ToUTF8String() optimizations
The modified c parameter of ToUTF8String is never used, so the "var" can be omitted. Moving the "Result" string to a var-parameter reduces the calls to LStrAsg (especially for attribute name and values) in the calling function, resulting in less CPU locks (Delphi).

This patch also adds a ToAsciiLowerUTF8String function that combines the ToUTF8String and TagNameToLower function. This removes unnecessary copying and modifying of a UTF8String.