Improved ParseNextNumEx performance and $OVERFLOWCHECKS fix
Changes the implementation of ParseNextNumEx to first parse the digits before the decimal separator with integer arithmetic with a fallback for very large numbers (int32->int64->double) and then parse the decimals also with integer arithmetic. After that the decimals are added to the result.
The Power(10, x) calls are replaced by pre-calculated constants if -18 <= x <= 18.