Improve x86 Trunc/Round code
The x86 FPU code is very slow when Trunc or Round is used. This patch adds the __Trunc method (that was already there in a previous version of Img32.pas) and adds a __Round function that, if the CPU supports SSE2 instructions, uses the SSE2 rounding code.

In order to not affect the 64 bit code, these two functions are mapped to Trunc and Round by using const function pointers only if "CPUX86" is defined. The Img32.pas probes the CPU via the CPUID instruction in the initialization block to detect SSE2 availability.