Improve x86 Trunc() performance
The x86 FPU code is very slow when Trunc is used. This patch adds the "__Trunc" function (that was already there in a previous version of Img32.pas).

In order to not affect the 64 bit code, the Trunc function is mapped to __Trunc by using a const function pointer that is only declared if "CPUX86" is defined.

The __Trunc function was formerly discussed in issue #45.