Fix for "code tidy" commit
DivOneByXTableSize = SizeOf(Word) results in an array with 2 instead of 65536 elements.
"array[Word]" means "array[0..65535]"

The last line in __Trunc applies the sign flag to "Result". The "if Value < 0" shouldn't be used here because it loads "Value" into the FPU what is slow and destroys the performance gain the function has. Instead a sign-bit test is used, so__Trunc can work without any FPU code.
The original PR used the TDoubleDataRec record. The new code uses an array[0..7] of Byte, making the function "inline" compatible.