TColorRenderer.RenderProc optimization
The optimized function only calls BlendToAlpha if the pixel is not identical to the previous or the alpha channel has changed. A special case handling for alpha=255 is also added (Rasterize).

In a test with "Koppen-Geiger_Map_Cfc_present.svg" the accumulated time for TColorRenderer.RenderProc went from over 3 seconds down to under 1 second.