Optimize FastGaussianBlur
If the whole image is to be blurred, the FastGaussianBlur function can now use the target image's pixels array directly instead of copying its contents into a helper array. This also means that the blurred image doesn't need to be copied back into the target image, because it is already there.

Another optimization is the use of CPU register optimized BoxBlurHLine/BoxBlurVLine helper functions.