Skip the use of the tmpImage if possible.
If the ShapeElement has a clip-element that is a rectangle and there are no other masks or filters and all paths are completely in the clip-region, it is possible to directly draw the shape into the target image instead of painting in a temporary image and the blending that into the target image. That improves the performance a lot if a SVG image contains many of such clip elements (e.g. Koppen-Geiger_Map_Cfc_present.svg).